add new functions js file
